Fixed num tassks completed metric
