+++ src/start-screen/components/HomePage.jsx
@@ -10,6 +10,8 @@ const SettingsButton = require("./SettingsButton");
 const {calcMetrics} = require("../../timeline/build-timeline.js");
 const data = require("../../timeline/two_hr_session.json"); //sample data
 const {convertTime, getTimeSpent} = require("../../timeline/calc-time.js");
+const {useTasks} = require("../../task_list/components/TaskContext");
+
 
 require("./../../timeline/components/navbarStyles.css");
 
@@ -18,11 +20,12 @@ const HomePage = () => {
   const [apiKey, setApiKey] = useState(localStorage.getItem('influxApiKey') || '');
   //Tracks if we've successfully connected to Influx
   const [dbSuccess, setDbSuccess] = useState(sessionStorage.getItem('db-status') || false);
+  const {numCompletedTasks} = useTasks();
 
   useEffect(() => {
     convertTime(data);
     getTimeSpent(data);
-    calcMetrics(data);
+    calcMetrics(data, 120, numCompletedTasks);
     ipcRenderer.on('db-conn-success', (_event) => {
       sessionStorage.setItem('db-status', true);
       setDbSuccess(true);
