+++ src/timeline/build-timeline.js
@@ -1,8 +1,6 @@
 /* eslint-disable no-console */
 
-const { getCompletedTasks } = require("../task_list/components/TaskContext");
-
-const calcMetrics = (data, sessionId) => {
+const calcMetrics = (data, duration, numTasksCompleted, sessionId) => {
   console.log("Calculating metrics for session ", sessionId);
   //Filling this with default values just to create a schema of our metrics
   const uniqueValues = [...new Set(data.map(item => item._value))];
@@ -13,13 +11,13 @@ const calcMetrics = (data, sessionId) => {
 
   const sessionMetrics = {
     //In reality, all of these except duration would be generated by a "process session ID" call of some kind
-    numTasks: getCompletedTasks(), // Number of tasks completed
-    tabSwitchRate: 0, // # of times tabs were switched divided by number of minutes in session
+    numTasks: numTasksCompleted, // Number of tasks completed
+    tabSwitchRate: data.length / duration, // # of times tabs were switched divided by number of minutes in session
     //timeOnDistr: 1000, // Secs of time spent on apps labeled as distractions, probably won't keep this
     //productivityEstimate: 8, // This isn't a real metric I just couldn't think of anything else
     mostUsedApps: sortedByTimeSpent.slice(0, 5).map(app => ({
       appName: app.name,
-      duration: app.totalTimeSpent,
+      durationInMinutes: app.totalTimeSpent,
       countSwitchedTo: app.numTimesSwitchedTo
     })),
     appSpecificMetrics: metricsByApp
